+++ CHANGELOG.rst
@@ -15,6 +15,10 @@ removed
 changes
 =======
 
+-  `WeldxFile` now hides ASDF added fields like history and asdf_library
+   from the dictionary interface. To access these, there are separate
+   properties `[#625] <https://github.com/BAMWeldX/weldx/pull/625>`__.
+
 fixes
 =====